split types into file
